files/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
#  JARVIS — Pinned Python Dependencies
#  Install: pip install -r requirements.txt
#
#  Versions pinned for reproducibility across Mac / Pi / PC nodes.
#  Run `pip install pip-compile-multi` and `pip-compile` to regenerate.
# ─────────────────────────────────────────────────────────────────────────────

# ── Config / env ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0          # .env file loading

# ── Core LLM + agent ──────────────────────────────────────────────────────────
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-ollama>=0.2.0,<0.3.0
langgraph>=0.2.0,<0.3.0

# ── LLM backends ─────────────────────────────────────────────────────────────
ollama>=0.3.0,<0.4.0                 # Ollama Python client (Mac / PC)

# ── Memory ────────────────────────────────────────────────────────────────────
chromadb>=0.5.0,<0.6.0
sentence-transformers>=3.0.0,<4.0.0  # fallback embeddings; also used on Pi

# ── PDF ingestion ─────────────────────────────────────────────────────────────
pymupdf>=1.24.0,<2.0.0               # fitz — fast PDF text + image extraction

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.12,<0.1.0      # file uploads

# ── Web / scraping ────────────────────────────────────────────────────────────
httpx>=0.27.0,<0.28.0
beautifulsoup4>=4.12.0,<5.0.0
ddgs>=6.3.0                          # renamed from duckduckgo-search
wikipedia-api>=0.7.0,<0.8.0         # for clipper.py

# ── Anthropic API (plugin writer) ─────────────────────────────────────────────
anthropic>=0.40.0,<0.41.0

# ── SSL / certs ───────────────────────────────────────────────────────────────
certifi>=2024.8.30

# ── macOS hotkey / menu bar ───────────────────────────────────────────────────
rumps>=0.4.0                         # menu bar app
pyobjc-framework-Cocoa>=10.0        # global hotkey support (macOS only)
pyobjc-framework-AppKit>=10.0

# ── Voice (uncomment when ready) ──────────────────────────────────────────────
# faster-whisper>=1.0.0,<2.0.0
# pyaudio>=0.2.14                    # brew install portaudio first

# ── Pi-specific (uncomment when deploying to Raspberry Pi) ───────────────────
# picamera2                           # Camera Module 3
# llama-cpp-python                    # llama.cpp bindings for Hailo

# ── PC / CUDA (uncomment for RTX node) ───────────────────────────────────────
# torch>=2.3.0                        # CUDA inference if needed
